@@ -1,59 +1,120 @@
 module github.com/ovh/venom
 
-go 1.16
+go 1.20
 
 require (
-	github.com/Azure/go-amqp v0.13.7
-	github.com/Shopify/sarama v1.27.2
-	github.com/alexbrainman/odbc v0.0.0-20200426075526-f0492dfa1575
-	github.com/antonfisher/nested-logrus-formatter v1.3.0
+	github.com/Azure/go-amqp v0.18.1
+	github.com/Shopify/sarama v1.38.1
+	github.com/alexbrainman/odbc v0.0.0-20211220213544-9c9a2e61c5e2
+	github.com/antonfisher/nested-logrus-formatter v1.3.1
 	github.com/confluentinc/bincover v0.2.0
-	github.com/eclipse/paho.mqtt.golang v1.3.4
-	github.com/fatih/color v1.10.0
-	github.com/fsamin/go-dump v1.2.0
-	github.com/fullstorydev/grpcurl v1.7.0
-	github.com/garyburd/redigo v1.6.2
-	github.com/ghodss/yaml v1.0.0
-	github.com/go-ole/go-ole v1.2.4 // indirect
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/go-testfixtures/testfixtures/v3 v3.4.1
-	github.com/golang/protobuf v1.4.3
-	github.com/golang/snappy v0.0.3 // indirect
+	github.com/eclipse/paho.mqtt.golang v1.4.2
+	github.com/fatih/color v1.15.0
+	github.com/fsamin/go-dump v1.8.0
+	github.com/fullstorydev/grpcurl v1.8.7
+	github.com/garyburd/redigo v1.6.4
+	github.com/go-sql-driver/mysql v1.7.0
+	github.com/go-testfixtures/testfixtures/v3 v3.8.1
+	github.com/golang/protobuf v1.5.3
 	github.com/google/go-github v17.0.0+incompatible
-	github.com/gosimple/slug v1.9.0
+	github.com/gosimple/slug v1.13.1
 	github.com/inconshreveable/go-update v0.0.0-20160112193335-8152e7eb6ccf
-	github.com/jhump/protoreflect v1.7.0
-	github.com/jmoiron/sqlx v1.2.0
+	github.com/jhump/protoreflect v1.15.1
+	github.com/jmoiron/sqlx v1.3.5
 	github.com/landoop/schema-registry v0.0.0-20190327143759-50a5701c1891
-	github.com/lib/pq v1.8.0
-	github.com/linkedin/goavro/v2 v2.10.0
-	github.com/mattn/go-shellwords v1.0.10
-	github.com/mattn/go-sqlite3 v2.0.3+incompatible // indirect
-	github.com/mattn/go-zglob v0.0.3
+	github.com/lib/pq v1.10.7
+	github.com/linkedin/goavro/v2 v2.12.0
+	github.com/mattn/go-shellwords v1.0.12
+	github.com/mattn/go-sqlite3 v2.0.3+incompatible
+	github.com/mattn/go-zglob v0.0.4
 	github.com/mitchellh/go-homedir v1.1.0
-	github.com/mitchellh/mapstructure v1.3.3
+	github.com/mitchellh/mapstructure v1.5.0
 	github.com/mndrix/tap-go v0.0.0-20171203230836-629fa407e90b
-	github.com/mxk/go-imap v0.0.0-20150429134902-531c36c3f12d // indirect
-	github.com/onsi/ginkgo v1.14.2 // indirect
-	github.com/onsi/gomega v1.10.3 // indirect
-	github.com/ovh/cds/sdk/interpolate v0.0.0-20201113080944-f9d9d58c157a
-	github.com/ovh/go-ovh v1.1.0
+	github.com/ovh/cds/sdk/interpolate v0.0.0-20230310144753-13590d1ea079
+	github.com/ovh/go-ovh v1.3.0
 	github.com/pkg/errors v0.9.1
 	github.com/r3labs/diff/v2 v2.14.0
 	github.com/rubenv/sql-migrate v0.0.0-20200616145509-8d140a17f351
-	github.com/sclevine/agouti v3.0.1-0.20180306165625-6ada53bb069e+incompatible
 	github.com/sijms/go-ora v0.0.0-20201108135513-712ea4f3d160
-	github.com/sirupsen/logrus v1.7.0
-	github.com/smartystreets/assertions v1.2.0 // indirect
-	github.com/spf13/cast v1.3.1
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/cast v1.4.1
 	github.com/spf13/cobra v1.1.1
 	github.com/spf13/pflag v1.0.5
 	github.com/streadway/amqp v1.0.0
-	github.com/stretchr/testify v1.6.1
+	github.com/stretchr/testify v1.8.2
 	github.com/yesnault/go-imap v0.0.0-20160710142244-eb9bbb66bd7b
-	golang.org/x/crypto v0.0.0-20201112155050-0c6587e931a9
-	golang.org/x/net v0.0.0-20201021035429-f5854403a974
-	golang.org/x/sys v0.0.0-20210320140829-1e4c9ba3b0c4 // indirect
-	google.golang.org/grpc v1.30.0
-	gopkg.in/yaml.v2 v2.3.0
+	go.mongodb.org/mongo-driver v1.11.2
+	golang.org/x/crypto v0.7.0
+	golang.org/x/exp v0.0.0-20230310171629-522b1b587ee0
+	golang.org/x/net v0.8.0
+	google.golang.org/grpc v1.53.0
+	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
+	modernc.org/sqlite v1.21.0
+)
+
+require (
+	github.com/aokoli/goutils v1.1.1 // indirect
+	github.com/bufbuild/protocompile v0.4.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/dustin/go-humanize v1.0.0 // indirect
+	github.com/eapache/go-resiliency v1.3.0 // indirect
+	github.com/eapache/go-xerial-snappy v0.0.0-20230111030713-bf00bc1b83b6 // indirect
+	github.com/eapache/queue v1.1.0 // indirect
+	github.com/golang/snappy v0.0.4 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/gosimple/unidecode v1.0.1 // indirect
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/hashicorp/go-uuid v1.0.3 // indirect
+	github.com/huandu/xstrings v1.4.0 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/jcmturner/aescts/v2 v2.0.0 // indirect
+	github.com/jcmturner/dnsutils/v2 v2.0.0 // indirect
+	github.com/jcmturner/gofork v1.7.6 // indirect
+	github.com/jcmturner/gokrb5/v8 v8.4.4 // indirect
+	github.com/jcmturner/rpc/v2 v2.0.3 // indirect
+	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
+	github.com/kevinramage/venomWeb v0.0.0-20230501110435-20012805e8a4
+	github.com/klauspost/compress v1.16.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
+	github.com/montanaflynn/stats v0.7.0 // indirect
+	github.com/mxk/go-imap v0.0.0-20150429134902-531c36c3f12d // indirect
+	github.com/pierrec/lz4/v4 v4.1.17 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rcrowley/go-metrics v0.0.0-20201227073835-cf1acfcdf475 // indirect
+	github.com/remyoudompheng/bigfft v0.0.0-20230129092748-24d4a6f8daec // indirect
+	github.com/xdg-go/pbkdf2 v1.0.0 // indirect
+	github.com/xdg-go/scram v1.1.2 // indirect
+	github.com/xdg-go/stringprep v1.0.4 // indirect
+	github.com/youmark/pkcs8 v0.0.0-20201027041543-1326539a0a0a // indirect
+	golang.org/x/mod v0.9.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/text v0.8.0 // indirect
+	golang.org/x/tools v0.7.0 // indirect
+	google.golang.org/genproto v0.0.0-20230306155012-7f2fa6fef1f4 // indirect
+	google.golang.org/protobuf v1.29.1 // indirect
+	gopkg.in/ini.v1 v1.67.0 // indirect
+	lukechampine.com/uint128 v1.2.0 // indirect
+	modernc.org/cc/v3 v3.40.0 // indirect
+	modernc.org/ccgo/v3 v3.16.13 // indirect
+	modernc.org/libc v1.22.3 // indirect
+	modernc.org/mathutil v1.5.0 // indirect
+	modernc.org/memory v1.5.0 // indirect
+	modernc.org/opt v0.1.3 // indirect
+	modernc.org/strutil v1.1.3 // indirect
+	modernc.org/token v1.1.0 // indirect
+)
+
+require github.com/rockbears/yaml v0.1.0
+
+require (
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/vmihailenco/msgpack v4.0.4+incompatible // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	gopkg.in/gorp.v1 v1.7.2 // indirect
 )
