@@ -10,8 +10,6 @@ import (
 	"time"
 
 	"github.com/mitchellh/mapstructure"
-	"github.com/pkg/errors"
-
 	"github.com/ovh/venom"
 )
 
@@ -38,8 +36,8 @@ type Executor struct {
 
 // Result represents a step result
 type Result struct {
-	Err         string  `json:"error,omitempty" yaml:"error"`
-	TimeSeconds float64 `json:"timeSeconds,omitempty" yaml:"timeSeconds,omitempty"`
+	Err         string  `json:"err,omitempty" yaml:"error"`
+	TimeSeconds float64 `json:"timeseconds,omitempty" yaml:"timeSeconds,omitempty"`
 }
 
 // ZeroValueResult return an empty implementation of this executor result
@@ -49,7 +47,7 @@ func (Executor) ZeroValueResult() interface{} {
 
 // GetDefaultAssertions return default assertions for type exec
 func (Executor) GetDefaultAssertions() *venom.StepAssertions {
-	return &venom.StepAssertions{Assertions: []string{"result.err ShouldBeEmpty"}}
+	return &venom.StepAssertions{Assertions: []venom.Assertion{"result.err ShouldBeEmpty"}}
 }
 
 // Run execute TestStep of type exec
@@ -96,6 +94,7 @@ func (e *Executor) sendEmail(ctx context.Context) error {
 	headers["From"] = e.From
 	headers["To"] = e.To
 	headers["Subject"] = e.Subject
+	headers["Content-Type"] = "text/plain; charset=utf-8"
 
 	// Setup message
 	message := ""
@@ -104,21 +103,19 @@ func (e *Executor) sendEmail(ctx context.Context) error {
 	}
 	message += "\r\n" + e.Body
 
-	// TLS config
 	tlsconfig := &tls.Config{
 		InsecureSkipVerify: true,
 		ServerName:         e.Host,
 	}
 
-	// Connect to the SMTP Server
 	servername := fmt.Sprintf("%s:%s", e.Host, e.Port)
-	venom.Info(ctx, "connecting to %s", servername)
+	venom.Info(ctx, "Connecting to %s", servername)
 
 	var c *smtp.Client
 	if e.WithTLS {
 		conn, err := tls.Dial("tcp", servername, tlsconfig)
 		if err != nil {
-			return errors.Wrapf(err, "tls dial error")
+			return fmt.Errorf("tls dial error: %w", err)
 		}
 
 		c, err = smtp.NewClient(conn, e.Host)
@@ -129,16 +126,15 @@ func (e *Executor) sendEmail(ctx context.Context) error {
 		var err error
 		c, err = smtp.Dial(servername)
 		if err != nil {
-			return errors.Wrapf(err, "tls dial error")
+			return fmt.Errorf("smtp dial error: %w", err)
 		}
 		defer c.Close()
 	}
 
-	// Auth
 	if e.User != "" && e.Password != "" {
 		auth := smtp.PlainAuth("", e.User, e.Password, e.Host)
 		if err := c.Auth(auth); err != nil {
-			return err
+			return fmt.Errorf("failed to authenticate: %w", err)
 		}
 	}
 
