@@ -1,25 +1,24 @@
 package ssh
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io/ioutil"
+	"github.com/mitchellh/mapstructure"
+	"github.com/ovh/venom"
+	"golang.org/x/crypto/ssh"
+	"io"
 	"os"
 	"os/user"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
-
-	"github.com/mitchellh/mapstructure"
-	"golang.org/x/crypto/ssh"
-
-	"github.com/ovh/venom"
+	"unicode/utf8"
 )
 
 // Name for test ssh
 const Name = "ssh"
+const sudoprompt = "sudo_venom"
 
 // New returns a new Test Exec
 func New() venom.Executor {
@@ -28,11 +27,13 @@ func New() venom.Executor {
 
 // Executor represents a Test Exec
 type Executor struct {
-	Host       string `json:"host,omitempty" yaml:"host,omitempty"`
-	Command    string `json:"command,omitempty" yaml:"command,omitempty"`
-	User       string `json:"user,omitempty" yaml:"user,omitempty"`
-	Password   string `json:"password,omitempty" yaml:"password,omitempty"`
-	PrivateKey string `json:"privatekey,omitempty" yaml:"privatekey,omitempty"`
+	Host         string `json:"host,omitempty" yaml:"host,omitempty"`
+	Command      string `json:"command,omitempty" yaml:"command,omitempty"`
+	User         string `json:"user,omitempty" yaml:"user,omitempty"`
+	Password     string `json:"password,omitempty" yaml:"password,omitempty"`
+	PrivateKey   string `json:"privatekey,omitempty" yaml:"privatekey,omitempty"`
+	Sudo         string `json:"sudo,omitempty" yaml:"sudo,omitempty"`
+	SudoPassword string `json:"sudopassword,omitempty" yaml:"sudopassword,omitempty"`
 }
 
 // Result represents a step result
@@ -51,7 +52,7 @@ func (Executor) ZeroValueResult() interface{} {
 
 // GetDefaultAssertions return default assertions for type exec
 func (Executor) GetDefaultAssertions() *venom.StepAssertions {
-	return &venom.StepAssertions{Assertions: []string{"result.code ShouldEqual 0"}}
+	return &venom.StepAssertions{Assertions: []venom.Assertion{"result.code ShouldEqual 0"}}
 }
 
 // Run execute TestStep of type exec
@@ -68,17 +69,30 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 	start := time.Now()
 	result := Result{}
 
-	client, session, err := connectToHost(e.User, e.Password, e.PrivateKey, e.Host)
+	client, session, err := connectToHost(e.User, e.Password, e.PrivateKey, e.Host, e.Sudo)
 	if err != nil {
 		result.Err = err.Error()
 	} else {
 		defer client.Close()
-		stdout := &bytes.Buffer{}
-		stderr := &bytes.Buffer{}
+		stdout := &Buffer{}
+		stderr := &Buffer{}
 
 		session.Stderr = stderr
 		session.Stdout = stdout
-		if err := session.Run(e.Command); err != nil {
+		stdin, _ := session.StdinPipe()
+
+		// Handle sudo password
+		command := e.Command
+		quit := make(chan bool)
+		if e.Sudo != "" {
+			command = "TERM=xterm-mono sudo -S -p " + sudoprompt + " -u " + e.Sudo + " " + command
+			if e.SudoPassword == "" {
+				e.SudoPassword = e.Password
+			}
+			go handleSudo(stdin, stdout, quit, e.SudoPassword)
+		}
+
+		if err := session.Run(command); err != nil {
 			if exiterr, ok := err.(*ssh.ExitError); ok {
 				status := exiterr.ExitStatus()
 				result.Code = strconv.Itoa(status)
@@ -93,8 +107,11 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 			result.Code = "0"
 		}
 
-		result.Systemerr = stderr.String()
-		result.Systemout = stdout.String()
+		if e.Sudo != "" {
+			quit <- true
+		}
+		result.Systemerr = strings.TrimSpace(stderr.String())
+		result.Systemout = strings.TrimSpace(stdout.String())
 	}
 
 	elapsed := time.Since(start)
@@ -103,7 +120,26 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 	return result, nil
 }
 
-func connectToHost(u, pass, key, host string) (*ssh.Client, *ssh.Session, error) {
+func handleSudo(in io.Writer, out *Buffer, quit chan bool, password string) {
+	sudopromptlen := len(sudoprompt)
+	for {
+		select {
+		case <-quit:
+			return
+		default:
+			content := out.String()
+			bufferLen := utf8.RuneCountInString(content)
+
+			// Check if we have to enter password
+			if bufferLen >= sudopromptlen && strings.Contains(content[bufferLen-sudopromptlen:], sudoprompt) {
+				in.Write([]byte(password + "\n"))
+				out.Truncate(0)
+			}
+		}
+	}
+}
+
+func connectToHost(u, pass, key, host, sudo string) (*ssh.Client, *ssh.Session, error) {
 	//Default user is current username
 	if u == "" {
 		osUser, err := user.Current()
@@ -113,9 +149,9 @@ func connectToHost(u, pass, key, host string) (*ssh.Client, *ssh.Session, error)
 		u = osUser.Username
 	}
 
-	//If password is set, use it
+	//If password is set, and we don't have key use it
 	var auth []ssh.AuthMethod
-	if pass != "" {
+	if pass != "" && key == "" {
 		auth = []ssh.AuthMethod{ssh.Password(pass)}
 	} else {
 		//Load the the private key
@@ -150,6 +186,18 @@ func connectToHost(u, pass, key, host string) (*ssh.Client, *ssh.Session, error)
 		return nil, nil, err
 	}
 
+	// Request PTY for sudo cmd
+	if sudo != "" {
+		modes := ssh.TerminalModes{
+			ssh.TTY_OP_ISPEED: 14400, // input speed = 14.4kbaud
+			ssh.TTY_OP_OSPEED: 14400, // output speed = 14.4kbaud
+		}
+
+		if err := session.RequestPty("xterm", 40, 80, modes); err != nil {
+			return nil, nil, err
+		}
+	}
+
 	return client, session, nil
 }
 
@@ -166,7 +214,7 @@ func privateKey(file string) (key ssh.Signer, err error) {
 	}
 
 	//Read the file
-	buf, err := ioutil.ReadFile(file)
+	buf, err := os.ReadFile(file)
 	if err != nil {
 		return nil, err
 	}
