@@ -3,15 +3,16 @@ package web
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"os"
 	"strings"
 	"time"
 
 	"github.com/gosimple/slug"
 	"github.com/mitchellh/mapstructure"
 	"github.com/pkg/errors"
-	"github.com/sclevine/agouti"
 
+	"github.com/kevinramage/venomWeb/common"
+	venomWeb "github.com/kevinramage/venomWeb/wrapper"
 	"github.com/ovh/venom"
 )
 
@@ -32,8 +33,8 @@ func New() venom.Executor {
 }
 
 type WebContext struct {
-	wd   *agouti.WebDriver
-	Page *agouti.Page
+	wd      venomWeb.WebDriver
+	session venomWeb.Session
 }
 
 // Executor struct
@@ -59,41 +60,54 @@ func (Executor) ZeroValueResult() interface{} {
 }
 
 func (Executor) Setup(ctx context.Context, vars venom.H) (context.Context, error) {
+	venom.Info(ctx, "Setup")
 	var webCtx WebContext
-	var driver = venom.StringVarFromCtx(ctx, "web.driver") // Possible values: chrome, phantomjs, gecko
-	var args = venom.StringVarFromCtx(ctx, "web.args")
+	var driver = venom.StringVarFromCtx(ctx, "web.driver") // Possible values: chrome, gecko
+	var args = venom.StringSliceVarFromCtx(ctx, "web.args")
 	var prefs = venom.StringMapInterfaceVarFromCtx(ctx, "web.prefs")
 
+	// Binary
+	var binaryPath = venom.StringVarFromCtx(ctx, "web.binaryPath")
+	var driverPath = venom.StringVarFromCtx(ctx, "web.driverPath")
+	var port = venom.StringVarFromCtx(ctx, "web.driverPort")
+
+	// Instanciate web driver (chrome by default)
 	switch driver {
-	case "chrome":
-		webCtx.wd = agouti.ChromeDriver(
-			agouti.ChromeOptions("args", args),
-			agouti.ChromeOptions("prefs", prefs),
-		)
 	case "gecko":
-		webCtx.wd = agouti.GeckoDriver()
+		webCtx.wd = venomWeb.GeckoDriver(binaryPath, driverPath, args, prefs, port)
 	default:
-		webCtx.wd = agouti.PhantomJS()
+		webCtx.wd = venomWeb.ChromeDriver(binaryPath, driverPath, args, prefs, port)
 	}
 
-	var timeout = venom.IntVarFromCtx(ctx, "web.timeout")
-	if timeout > 0 {
+	// Configure web driver
+	if timeout := venom.IntVarFromCtx(ctx, "web.timeout"); timeout > 0 {
 		webCtx.wd.Timeout = time.Duration(timeout) * time.Second
-	} else {
-		webCtx.wd.Timeout = 180 * time.Second // default value
+	}
+	if debug := venom.BoolVarFromCtx(ctx, "web.debug"); debug {
+		webCtx.wd.LogLevel = common.DEBUG
+	}
+	if proxy := venom.StringVarFromCtx(ctx, "web.proxy"); proxy != "" {
+		webCtx.wd.Proxy = proxy
+	}
+	if headless := venom.BoolVarFromCtx(ctx, "web.headless"); headless {
+		webCtx.wd.Headless = headless
+	}
+	if detach := venom.BoolVarFromCtx(ctx, "web.detach"); detach {
+		webCtx.wd.Detach = detach
+	}
+	if logLevel := venom.StringVarFromCtx(ctx, "web.logLevel"); logLevel != "" {
+		webCtx.wd.LogLevel = logLevel
 	}
 
-	webCtx.wd.Debug = venom.BoolVarFromCtx(ctx, "web.debug")
-
-	if err := webCtx.wd.Start(); err != nil {
-		return ctx, errors.Wrapf(err, "Unable start web driver")
-	}
-
-	// Init Page
+	// Init session
 	var err error
-	webCtx.Page, err = webCtx.wd.NewPage()
+	err = webCtx.wd.Start()
 	if err != nil {
-		return ctx, errors.Wrapf(err, "Unable create new page")
+		return ctx, errors.Wrapf(err, "Unable to start web driver")
+	}
+	webCtx.session, err = webCtx.wd.NewSession()
+	if err != nil {
+		return ctx, errors.Wrapf(err, "Unable create new session")
 	}
 
 	var resizePage bool
@@ -105,8 +119,8 @@ func (Executor) Setup(ctx context.Context, vars venom.H) (context.Context, error
 
 	// Resize Page
 	if resizePage {
-		if err := webCtx.Page.Size(width, height); err != nil {
-			return ctx, fmt.Errorf("Unable resize page: %s", err)
+		if err := webCtx.session.Size(width, height); err != nil {
+			return ctx, fmt.Errorf("unable resize page: %s", err)
 		}
 	}
 
@@ -122,6 +136,7 @@ func getWebCtx(ctx context.Context) *WebContext {
 }
 
 func (Executor) TearDown(ctx context.Context) error {
+	venom.Info(ctx, "TearDown")
 	return getWebCtx(ctx).wd.Stop()
 }
 
@@ -137,10 +152,10 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 		return nil, err
 	}
 
-	result, err := e.runAction(ctx, webCtx.Page)
+	result, err := e.runAction(ctx, webCtx.session)
 	if err != nil {
-		if errg := generateErrorHTMLFile(ctx, webCtx.Page, slug.Make(webCtx.Page.String())); errg != nil {
-			venom.Warn(ctx, "Error while generate HTML file: %v", errg)
+		if errg := generateErrorHTMLFile(ctx, webCtx.session, slug.Make(webCtx.session.String())); errg != nil {
+			venom.Warn(ctx, "Error while generating the HTML file: %v", errg)
 			return nil, err
 		}
 		return nil, err
@@ -148,18 +163,19 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 
 	// take a screenshot
 	if e.Screenshot != "" {
-		if err := webCtx.Page.Screenshot(e.Screenshot); err != nil {
+		if err := webCtx.session.TakeScreenshot(e.Screenshot); err != nil {
 			return nil, err
 		}
-		if err := generateErrorHTMLFile(ctx, webCtx.Page, slug.Make(webCtx.Page.String())); err != nil {
-			venom.Warn(ctx, "Error while generate HTML file: %v", err)
+		if err := generateErrorHTMLFile(ctx, webCtx.session, slug.Make(webCtx.session.String())); err != nil {
+			venom.Warn(ctx, "Error while generating the HTML file: %v", err)
 			return nil, err
 		}
 	}
 
 	// Get page title (Check the absence of popup before the page title collect to avoid error)
-	if _, err := webCtx.Page.PopupText(); err != nil {
-		title, err := webCtx.Page.Title()
+	result.Title = ""
+	if _, err := webCtx.session.GetAlertText(); err != nil {
+		title, err := webCtx.session.GetTitle()
 		if err != nil {
 			return nil, err
 		}
@@ -167,10 +183,11 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 	}
 
 	// Get page url (Check the absence of popup before the page url collect to avoid error)
-	if _, err := webCtx.Page.PopupText(); err != nil {
-		url, errU := webCtx.Page.URL()
-		if errU != nil {
-			return nil, fmt.Errorf("Cannot get URL: %s", errU)
+	result.URL = ""
+	if _, err := webCtx.session.GetAlertText(); err != nil {
+		url, err := webCtx.session.GetURL()
+		if err != nil {
+			return nil, fmt.Errorf("cannot get url: %s", err)
 		}
 		result.URL = url
 	}
@@ -181,175 +198,241 @@ func (Executor) Run(ctx context.Context, step venom.TestStep) (interface{}, erro
 	return result, nil
 }
 
-func (e Executor) runAction(ctx context.Context, page *agouti.Page) (*Result, error) {
+func (e Executor) runAction(ctx context.Context, session venomWeb.Session) (*Result, error) {
 	r := &Result{}
+
+	// Click
 	if e.Action.Click != nil {
-		s, err := find(page, e.Action.Click.Find, r)
+
+		// Find element
+		elt, err := findOne(ctx, session, e.Action.Click.Find, e.Action.Click.SyncTimeout)
 		if err != nil {
 			return nil, err
 		}
-		if err := s.Click(); err != nil {
+
+		// Click on element
+		if err := elt.Click(); err != nil {
 			return nil, err
 		}
+
+		// Wait
 		if e.Action.Click.Wait != 0 {
 			time.Sleep(time.Duration(e.Action.Click.Wait) * time.Second)
 		}
+
+		// Fill
 	} else if e.Action.Fill != nil {
 		for _, f := range e.Action.Fill {
-			s, err := findOne(page, f.Find, r)
+			elt, err := findOne(ctx, session, f.Find, f.SyncTimeout)
 			if err != nil {
 				return nil, err
 			}
-			if err := s.Fill(f.Text); err != nil {
+			if err := elt.SendKeys(f.Text); err != nil {
 				return nil, err
 			}
 			if f.Key != nil {
-				if err := s.SendKeys(Keys[*f.Key]); err != nil {
+				if err := elt.SendKeys(Keys[*f.Key]); err != nil {
 					return nil, err
 				}
 			}
 		}
-	} else if e.Action.Find != "" {
-		s, err := find(page, e.Action.Find, r)
+	} else if e.Action.Find != nil {
+		elts, err := find(ctx, session, e.Action.Find, r)
 		if err != nil {
 			return nil, err
-		} else if s != nil {
-			if count, errCount := s.Count(); errCount == nil && count == 1 {
-				if elements, errElements := s.Elements(); errElements == nil {
-					if text, errText := elements[0].GetText(); errText == nil {
-						r.Text = text
-					}
-					if value, errValue := elements[0].GetAttribute("value"); errValue == nil {
-						r.Value = value
-					}
-				}
+		} else if len(elts) > 0 {
+			if text, err := elts[0].GetElementText(); err == nil {
+				r.Text = text
+			}
+			if value, err := elts[0].GetElementProperty("value"); err == nil {
+				r.Value = value
 			}
 		}
+
+		// Navigate
 	} else if e.Action.Navigate != nil {
-		if err := page.Navigate(e.Action.Navigate.URL); err != nil {
+		if e.Action.Navigate.Reset {
+			if err := session.Reset(); err != nil {
+				return nil, err
+			}
+		}
+		if err := session.Navigate(e.Action.Navigate.URL); err != nil {
 			return nil, err
 		}
-		if e.Action.Navigate.Reset {
-			if err := page.Reset(); err != nil {
-				return nil, err
-			}
-			if err := page.Navigate(e.Action.Navigate.URL); err != nil {
-				return nil, err
-			}
-		}
+
+		// Wait
 	} else if e.Action.Wait != 0 {
 		time.Sleep(time.Duration(e.Action.Wait) * time.Second)
+
+		// Confirm popup
 	} else if e.Action.ConfirmPopup {
-		if err := page.ConfirmPopup(); err != nil {
+		if err := session.AcceptAlert(); err != nil {
 			return nil, err
 		}
+
+		// Cancel popup
 	} else if e.Action.CancelPopup {
-		if err := page.CancelPopup(); err != nil {
+		if err := session.DismissAlert(); err != nil {
 			return nil, err
 		}
+
+		// Select
 	} else if e.Action.Select != nil {
-		s, err := findOne(page, e.Action.Select.Find, r)
+		elt, err := findOne(ctx, session, e.Action.Select.Find, e.Action.Select.SyncTimeout)
 		if err != nil {
 			return nil, err
 		}
-		if err := s.Select(e.Action.Select.Text); err != nil {
+		if _, err := elt.Select(e.Action.Select.Text); err != nil {
 			return nil, err
 		}
 		if e.Action.Select.Wait != 0 {
 			time.Sleep(time.Duration(e.Action.Select.Wait) * time.Second)
 		}
+
+		// Upload file
 	} else if e.Action.UploadFile != nil {
-		s, err := find(page, e.Action.UploadFile.Find, r)
+		elt, err := findOne(ctx, session, e.Action.UploadFile.Find, e.Action.UploadFile.SyncTimeout)
 		if err != nil {
 			return nil, err
 		}
 		for _, f := range e.Action.UploadFile.Files {
-			if err := s.UploadFile(f); err != nil {
+			if err := elt.UploadFile(f); err != nil {
 				return nil, err
 			}
 		}
 		if e.Action.UploadFile.Wait != 0 {
 			time.Sleep(time.Duration(e.Action.UploadFile.Wait) * time.Second)
 		}
+
+		// Select frame
 	} else if e.Action.SelectFrame != nil {
-		s, err := findOne(page, e.Action.SelectFrame.Find, r)
+		elt, err := findOne(ctx, session, e.Action.SelectFrame.Find, e.Action.SelectFrame.SyncTimeout)
 		if err != nil {
 			return nil, err
 		}
-		if elements, errElements := s.Elements(); errElements == nil {
-			if errSelectFrame := page.Session().Frame(elements[0]); errSelectFrame != nil {
-				return nil, errSelectFrame
-			}
-		} else {
-			return nil, errElements
+		err = session.SwitchToFrame(elt)
+		if err != nil {
+			return nil, err
 		}
+
+		// Select root frame
 	} else if e.Action.SelectRootFrame {
-		if err := page.SwitchToRootFrame(); err != nil {
+		if err := session.SwitchToParentFrame(); err != nil {
 			return nil, err
 		}
+
+		// Next window
 	} else if e.Action.NextWindow {
-		if err := page.NextWindow(); err != nil {
+		if err := session.NextWindow(); err != nil {
 			return nil, err
 		}
+
+		// Back, Forward, Refresh
 	} else if e.Action.HistoryAction != "" {
 		switch strings.ToLower(e.Action.HistoryAction) {
 		case "back":
-			if err := page.Back(); err != nil {
+			if err := session.Back(); err != nil {
 				return nil, err
 			}
 		case "refresh":
-			if err := page.Refresh(); err != nil {
+			if err := session.Refresh(); err != nil {
 				return nil, err
 			}
 		case "forward":
-			if err := page.Forward(); err != nil {
+			if err := session.Forward(); err != nil {
 				return nil, err
 			}
 		default:
-			return nil, fmt.Errorf("History action '%s' is invalid", e.Action.HistoryAction)
+			return nil, fmt.Errorf("history action '%s' is invalid", e.Action.HistoryAction)
+		}
+
+		// Execute
+	} else if e.Action.Execute != nil {
+		args := []string{}
+		if e.Action.Execute.Args != nil && len(e.Action.Execute.Args) > 0 {
+			args = e.Action.Execute.Args
+		}
+		err := session.ExecuteScript(e.Action.Execute.Command, args)
+		if err != nil {
+			return nil, err
 		}
 	}
+
 	return r, nil
 }
 
-func find(page *agouti.Page, search string, r *Result) (*agouti.Selection, error) {
-	s := page.Find(search)
-	if s == nil {
-		return nil, fmt.Errorf("Cannot find element %s", search)
+func find(ctx context.Context, session venomWeb.Session, findElement interface{}, r *Result) ([]venomWeb.Element, error) {
+
+	// Identify selector and locator strategy
+	selector, locator, err := readFindElement(ctx, findElement)
+	if err != nil {
+		return nil, err
 	}
-	nbElement, errC := s.Count()
-	if errC != nil {
-		if !strings.Contains(errC.Error(), "element not found") {
-			return nil, fmt.Errorf("Cannot count element %s: %s", search, errC)
-		}
-		nbElement = 0
+
+	// Search elements
+	elts, err := session.FindElements(selector, locator)
+	if err != nil {
+		return nil, err
 	}
-	r.Find = nbElement
-	return s, nil
+	r.Find = len(elts)
+
+	// Update result
+	if r.Find == 1 {
+		r.Text, _ = elts[0].GetElementText()
+		r.Value, _ = elts[0].GetElementProperty("value")
+	}
+
+	return elts, nil
 }
 
-func findOne(page *agouti.Page, search string, r *Result) (*agouti.Selection, error) {
-	s := page.Find(search)
-	if s == nil {
-		return nil, fmt.Errorf("Cannot find element %s", search)
+// Find element from a selector
+func findOne(ctx context.Context, session venomWeb.Session, findElement interface{}, syncTimeout int64) (venomWeb.Element, error) {
+
+	// Identify selector and locator strategy
+	selector, locator, err := readFindElement(ctx, findElement)
+	if err != nil {
+		return venomWeb.Element{}, err
 	}
-	nbElement, errC := s.Count()
-	if errC != nil {
-		return nil, fmt.Errorf("Cannot find element %s: %s", search, errC)
+
+	// Synchronize element
+	if syncTimeout > 0 {
+		if err := session.SyncElement(selector, locator, syncTimeout*1000); err != nil {
+			return venomWeb.Element{}, err
+		}
 	}
-	if nbElement != 1 {
-		return nil, fmt.Errorf("Find %d elements", nbElement)
+
+	// Find element
+	return session.FindElement(selector, locator)
+}
+
+// Identify selector and locator from a find element
+func readFindElement(ctx context.Context, findElement interface{}) (string, string, error) {
+	selector := ""
+	locator := common.CSS_SELECTOR
+	if find, ok := findElement.(string); ok {
+		venom.Warning(ctx, "web - findElement : this find element syntax deprecated and will be not supported in next version")
+		selector = find
+	} else if find, ok := findElement.(map[string]interface{}); ok {
+		selector = find["selector"].(string)
+		if find["locator"] == "XPATH" {
+			locator = common.XPATH_SELECTOR
+		} else if find["locator"] != "CSS" {
+			return "", "", fmt.Errorf("invalid find element locator '%s' (must be CSS or XPATH)", find["locator"])
+		}
+	} else {
+		return "", "", fmt.Errorf("invalid find element structure %v\n (must be a string or a element with selector and locator attribute)", findElement)
 	}
-	return s, nil
+
+	return selector, locator, nil
 }
 
 // generateErrorHTMLFile generates an HTML file in error case to identify clearly the error
-func generateErrorHTMLFile(ctx context.Context, page *agouti.Page, name string) error {
-	html, err := page.HTML()
+func generateErrorHTMLFile(ctx context.Context, session venomWeb.Session, name string) error {
+	html, err := session.GetPageSource()
 	if err != nil {
 		return err
 	}
 	filename := name + ".dump.html"
 	venom.Info(ctx, "Content of the HTML page is saved in %s", filename)
-	return ioutil.WriteFile(filename, []byte(html), 0644)
+	return os.WriteFile(filename, []byte(html), 0644)
 }
