@@ -3,14 +3,12 @@ package venom
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"time"
 
-	"github.com/ghodss/yaml"
+	"github.com/rockbears/yaml"
 
 	"github.com/mattn/go-zglob"
 
@@ -27,7 +25,8 @@ func getFilesPath(path []string) (filePaths []string, err error) {
 		// if we put ./test/*.yml, it will fail and it's normal
 		fileInfo, _ := os.Stat(p)
 		if fileInfo != nil && fileInfo.IsDir() {
-			p = p + string(os.PathSeparator) + "*.yml"
+			//check if *.yml or *.yaml files exists in the path
+			p = p + string(os.PathSeparator) + "*.y*ml"
 		}
 
 		fpaths, err := zglob.Glob(p)
@@ -44,9 +43,8 @@ func getFilesPath(path []string) (filePaths []string, err error) {
 		}
 	}
 
-	sort.Strings(filePaths)
 	if len(filePaths) == 0 {
-		return nil, fmt.Errorf("no yml file selected")
+		return nil, fmt.Errorf("no YAML (*.yml or *.yaml) file found or defined")
 	}
 	return uniq(filePaths), nil
 }
@@ -69,40 +67,49 @@ type partialTestSuite struct {
 }
 
 func (v *Venom) readFiles(ctx context.Context, filesPath []string) (err error) {
-	for _, f := range filesPath {
-		log.Info("Reading ", f)
-		btes, err := ioutil.ReadFile(f)
+	for _, filePath := range filesPath {
+		log.Info("Reading ", filePath)
+		btes, err := os.ReadFile(filePath)
 		if err != nil {
-			return errors.Wrapf(err, "unable to read file %q", f)
+			return errors.Wrapf(err, "unable to read file %q", filePath)
 		}
 
 		varCloned := v.variables.Clone()
 
 		fromPartial, err := getVarFromPartialYML(ctx, btes)
 		if err != nil {
-			return errors.Wrapf(err, "unable to get vars from file %q", f)
+			return errors.Wrapf(err, "unable to get vars from file %q", filePath)
 		}
 
-		varsFromPartial, err := DumpStringPreserveCase(fromPartial)
-		if err != nil {
-			return errors.Wrapf(err, "unable to parse variables")
+		var varsFromPartial map[string]string
+		if len(fromPartial) > 0 {
+			varsFromPartial, err = DumpStringPreserveCase(fromPartial)
+			if err != nil {
+				return errors.Wrapf(err, "unable to parse variables")
+			}
 		}
 
 		// we take default vars from the testsuite, only if it's not already is global vars
 		for k, value := range varsFromPartial {
+			if k == "" {
+				continue
+			}
 			if _, ok := varCloned[k]; !ok || (varCloned[k] == "{}" && varCloned["__Len__"] == "0") {
 				// we interpolate the value of vars here, to do it only once per ts
 				valueInterpolated, err := interpolate.Do(value, varsFromPartial)
 				if err != nil {
 					return errors.Wrapf(err, "unable to parse variable %q", k)
 				}
-				varCloned[k] = valueInterpolated
+				varCloned.Add(k, valueInterpolated)
 			}
 		}
 
-		vars, err := DumpStringPreserveCase(varCloned)
-		if err != nil {
-			return errors.Wrapf(err, "unable to parse variables")
+		var vars map[string]string
+		if len(varCloned) > 0 {
+			vars, err = DumpStringPreserveCase(varCloned)
+			if err != nil {
+				return errors.Wrapf(err, "unable to parse variables")
+			}
 		}
 
 		content, err := interpolate.Do(string(btes), vars)
@@ -110,38 +117,46 @@ func (v *Venom) readFiles(ctx context.Context, filesPath []string) (err error) {
 			return err
 		}
 
-		var ts TestSuite
-		if err := yaml.Unmarshal([]byte(content), &ts); err != nil {
+		var testSuiteInput TestSuiteInput
+		if err := yaml.Unmarshal([]byte(content), &testSuiteInput); err != nil {
 			Error(context.Background(), "file content: %s", content)
-			return errors.Wrapf(err, "error while unmarshal file %q", f)
+			return errors.Wrapf(err, "error while unmarshal file %q", filePath)
+		}
+
+		ts := TestSuite{
+			Name:      testSuiteInput.Name,
+			TestCases: make([]TestCase, len(testSuiteInput.TestCases)),
+			Vars:      testSuiteInput.Vars,
+		}
+		for i := range testSuiteInput.TestCases {
+			ts.TestCases[i] = TestCase{
+				TestCaseInput: testSuiteInput.TestCases[i],
+			}
 		}
 
 		// Default workdir is testsuite directory
-		ts.WorkDir, err = filepath.Abs(filepath.Dir(f))
+		ts.WorkDir, err = filepath.Abs(filepath.Dir(filePath))
 		if err != nil {
 			return errors.Wrapf(err, "Unable to get testsuite's working directory")
 		}
 
-		ts.Package = f
-		ts.Filename = f
+		// ../foo/a.yml
+		ts.Filepath = filePath
+		// a
+		ts.ShortName = strings.TrimSuffix(filepath.Base(filePath), filepath.Ext(filePath))
+		// a.yml
+		ts.Filename = filepath.Base(filePath)
 		ts.Vars = varCloned
 
 		ts.Vars.Add("venom.testsuite.workdir", ts.WorkDir)
-		ts.Vars.Add("venom.testsuite.shortName", ts.Name)
+		ts.Vars.Add("venom.testsuite.name", ts.Name)
+		ts.Vars.Add("venom.testsuite.shortName", ts.ShortName)
 		ts.Vars.Add("venom.testsuite.filename", ts.Filename)
+		ts.Vars.Add("venom.testsuite.filepath", ts.Filepath)
 		ts.Vars.Add("venom.datetime", time.Now().Format(time.RFC3339))
 		ts.Vars.Add("venom.timestamp", fmt.Sprintf("%d", time.Now().Unix()))
 
-		nSteps := 0
-		for _, tc := range ts.TestCases {
-			nSteps += len(tc.testSteps)
-			if len(tc.Skipped) >= 1 {
-				ts.Skipped += len(tc.Skipped)
-			}
-		}
-		ts.Total = len(ts.TestCases)
-
-		v.testsuites = append(v.testsuites, ts)
+		v.Tests.TestSuites = append(v.Tests.TestSuites, ts)
 	}
 	return nil
 }
