@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -14,10 +13,11 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/confluentinc/bincover"
 	"github.com/fatih/color"
-	"github.com/ghodss/yaml"
 	"github.com/ovh/cds/sdk/interpolate"
 	"github.com/pkg/errors"
+	"github.com/rockbears/yaml"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cast"
 )
@@ -25,21 +25,31 @@ import (
 var (
 	//Version is set with -ldflags "-X github.com/ovh/venom/venom.Version=$(VERSION)"
 	Version = "snapshot"
+	IsTest  = ""
 )
+
+func OSExit(exitCode int) {
+	if IsTest != "" {
+		bincover.ExitCode = exitCode
+	} else {
+		os.Exit(exitCode)
+	}
+}
 
 // ContextKey can be added in context to store contextual infos. Also used by logger.
 type ContextKey string
 
-// New instanciates a new venom on venom run cmd
+// New instantiates a new venom on venom run cmd
 func New() *Venom {
 	v := &Venom{
-		LogOutput:        os.Stdout,
-		PrintFunc:        fmt.Printf,
-		executorsBuiltin: map[string]Executor{},
-		executorsPlugin:  map[string]Executor{},
-		executorsUser:    map[string]Executor{},
-		variables:        map[string]interface{}{},
-		OutputFormat:     "xml",
+		LogOutput:         os.Stdout,
+		PrintFunc:         fmt.Printf,
+		executorsBuiltin:  map[string]Executor{},
+		executorsPlugin:   map[string]Executor{},
+		executorsUser:     map[string]Executor{},
+		executorFileCache: map[string][]byte{},
+		variables:         map[string]interface{}{},
+		OutputFormat:      "xml",
 	}
 	return v
 }
@@ -47,18 +57,20 @@ func New() *Venom {
 type Venom struct {
 	LogOutput io.Writer
 
-	PrintFunc        func(format string, a ...interface{}) (n int, err error)
-	executorsBuiltin map[string]Executor
-	executorsPlugin  map[string]Executor
-	executorsUser    map[string]Executor
+	PrintFunc         func(format string, a ...interface{}) (n int, err error)
+	executorsBuiltin  map[string]Executor
+	executorsPlugin   map[string]Executor
+	executorsUser     map[string]Executor
+	executorFileCache map[string][]byte
 
-	testsuites []TestSuite
-	variables  H
+	Tests     Tests
+	variables H
 
 	LibDir        string
 	OutputFormat  string
 	OutputDir     string
 	StopOnFailure bool
+	HtmlReport    bool
 	Verbose       int
 }
 
@@ -73,7 +85,11 @@ func (v *Venom) Println(format string, a ...interface{}) {
 }
 
 func (v *Venom) PrintlnTrace(s string) {
-	v.Println("\t  %s %s", trace("[trac]"), trace(s)) // nolint
+	v.PrintlnIndentedTrace(s, "")
+}
+
+func (v *Venom) PrintlnIndentedTrace(s string, indent string) {
+	v.Println("\t  %s%s %s", indent, trace("[trac]"), trace(s)) // nolint
 }
 
 func (v *Venom) AddVariables(variables map[string]interface{}) {
@@ -101,10 +117,15 @@ func (v *Venom) RegisterExecutorUser(name string, e Executor) {
 // no type -> exec is default
 func (v *Venom) GetExecutorRunner(ctx context.Context, ts TestStep, h H) (context.Context, ExecutorRunner, error) {
 	name, _ := ts.StringValue("type")
-	if name == "" {
+	script, _ := ts.StringValue("script")
+	if name == "" && script != "" {
 		name = "exec"
 	}
 	retry, err := ts.IntValue("retry")
+	if err != nil {
+		return nil, nil, err
+	}
+	retryIf, err := ts.StringSliceValue("retry_if")
 	if err != nil {
 		return nil, nil, err
 	}
@@ -130,16 +151,20 @@ func (v *Venom) GetExecutorRunner(ctx context.Context, ts TestStep, h H) (contex
 	}
 	ctx = context.WithValue(ctx, ContextKey("vars"), allKeys)
 
-	if ex, ok := v.executorsBuiltin[name]; ok {
-		return ctx, newExecutorRunner(ex, name, "builtin", retry, delay, timeout, info), nil
+	if name == "" {
+		return ctx, newExecutorRunner(nil, name, "builtin", retry, retryIf, delay, timeout, info), nil
 	}
 
-	if err := v.registerUserExecutors(ctx, name, ts, vars); err != nil {
+	if ex, ok := v.executorsBuiltin[name]; ok {
+		return ctx, newExecutorRunner(ex, name, "builtin", retry, retryIf, delay, timeout, info), nil
+	}
+
+	if err := v.registerUserExecutors(ctx, name, vars); err != nil {
 		Debug(ctx, "executor %q is not implemented as user executor - err:%v", name, err)
 	}
 
 	if ex, ok := v.executorsUser[name]; ok {
-		return ctx, newExecutorRunner(ex, name, "user", retry, delay, timeout, info), nil
+		return ctx, newExecutorRunner(ex, name, "user", retry, retryIf, delay, timeout, info), nil
 	}
 
 	if err := v.registerPlugin(ctx, name, vars); err != nil {
@@ -148,7 +173,7 @@ func (v *Venom) GetExecutorRunner(ctx context.Context, ts TestStep, h H) (contex
 
 	// then add the executor plugin to the map to not have to load it on each step
 	if ex, ok := v.executorsUser[name]; ok {
-		return ctx, newExecutorRunner(ex, name, "plugin", retry, delay, timeout, info), nil
+		return ctx, newExecutorRunner(ex, name, "plugin", retry, retryIf, delay, timeout, info), nil
 	}
 	return ctx, nil, fmt.Errorf("executor %q is not implemented", name)
 }
@@ -156,9 +181,8 @@ func (v *Venom) GetExecutorRunner(ctx context.Context, ts TestStep, h H) (contex
 func (v *Venom) getUserExecutorFilesPath(vars map[string]string) (filePaths []string, err error) {
 	var libpaths []string
 	if v.LibDir != "" {
-		for _, p := range strings.Split(v.LibDir, string(os.PathListSeparator)) {
-			libpaths = append(libpaths, p)
-		}
+		p := strings.Split(v.LibDir, string(os.PathListSeparator))
+		libpaths = append(libpaths, p...)
 	}
 	libpaths = append(libpaths, path.Join(vars["venom.testsuite.workdir"], "lib"))
 
@@ -172,6 +196,9 @@ func (v *Venom) getUserExecutorFilesPath(vars map[string]string) (filePaths []st
 			}
 			return nil
 		})
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	sort.Strings(filePaths)
@@ -181,7 +208,7 @@ func (v *Venom) getUserExecutorFilesPath(vars map[string]string) (filePaths []st
 	return filePaths, nil
 }
 
-func (v *Venom) registerUserExecutors(ctx context.Context, name string, ts TestStep, vars map[string]string) error {
+func (v *Venom) registerUserExecutors(ctx context.Context, name string, vars map[string]string) error {
 	executorsPath, err := v.getUserExecutorFilesPath(vars)
 	if err != nil {
 		return err
@@ -189,9 +216,13 @@ func (v *Venom) registerUserExecutors(ctx context.Context, name string, ts TestS
 
 	for _, f := range executorsPath {
 		log.Info("Reading ", f)
-		btes, err := ioutil.ReadFile(f)
-		if err != nil {
-			return errors.Wrapf(err, "unable to read file %q", f)
+		btes, ok := v.executorFileCache[f]
+		if !ok {
+			btes, err = os.ReadFile(f)
+			if err != nil {
+				return errors.Wrapf(err, "unable to read file %q", f)
+			}
+			v.executorFileCache[f] = btes
 		}
 
 		varsFromInput, err := getUserExecutorInputYML(ctx, btes)
@@ -200,9 +231,12 @@ func (v *Venom) registerUserExecutors(ctx context.Context, name string, ts TestS
 		}
 
 		// varsFromInput contains the default vars from the executor
-		varsFromInputMap, err := DumpStringPreserveCase(varsFromInput)
-		if err != nil {
-			return errors.Wrapf(err, "unable to parse variables")
+		var varsFromInputMap map[string]string
+		if len(varsFromInput) > 0 {
+			varsFromInputMap, err = DumpStringPreserveCase(varsFromInput)
+			if err != nil {
+				return errors.Wrapf(err, "unable to parse variables")
+			}
 		}
 
 		varsComputed := map[string]string{}
@@ -225,6 +259,8 @@ func (v *Venom) registerUserExecutors(ctx context.Context, name string, ts TestS
 		if err := yaml.Unmarshal([]byte(content), &ux); err != nil {
 			return errors.Wrapf(err, "unable to parse file %q with content %v", f, content)
 		}
+
+		log.Debugf("User executor %q revolved with content %v", f, content)
 
 		for k, vr := range varsComputed {
 			ux.Input.Add(k, vr)
