@@ -16,27 +16,29 @@ import (
 	"github.com/ovh/venom/assertions"
 )
 
-type assertionsApplied struct {
-	ok        bool
-	errors    []Failure
-	failures  []Failure
-	systemout string
-	systemerr string
+type AssertionsApplied struct {
+	OK         bool `json:"ok" yml:"-"`
+	errors     []Failure
+	systemout  string
+	systemerr  string
+	Assertions []AssertionApplied `json:"assertions" yml:"-"`
+}
+type AssertionApplied struct {
+	Assertion Assertion `json:"assertion" yml:"-"`
+	IsOK      bool      `json:"isOK" yml:"-"`
 }
 
-func applyAssertions(r interface{}, tc TestCase, stepNumber int, step TestStep, defaultAssertions *StepAssertions) assertionsApplied {
+func applyAssertions(ctx context.Context, r interface{}, tc TestCase, stepNumber int, rangedIndex int, step TestStep, defaultAssertions *StepAssertions) AssertionsApplied {
 	var sa StepAssertions
 	var errors []Failure
-	var failures []Failure
 	var systemerr, systemout string
 
 	if err := mapstructure.Decode(step, &sa); err != nil {
-		return assertionsApplied{
-			false,
-			[]Failure{{Value: RemoveNotPrintableChar(fmt.Sprintf("error decoding assertions: %s", err))}},
-			failures,
-			systemout,
-			systemerr,
+		return AssertionsApplied{
+			OK:        false,
+			errors:    []Failure{{Value: RemoveNotPrintableChar(fmt.Sprintf("error decoding assertions: %s", err))}},
+			systemout: systemout,
+			systemerr: systemerr,
 		}
 	}
 
@@ -47,16 +49,19 @@ func applyAssertions(r interface{}, tc TestCase, stepNumber int, step TestStep, 
 	executorResult := GetExecutorResult(r)
 
 	isOK := true
+	assertions := []AssertionApplied{}
 	for _, assertion := range sa.Assertions {
-		errs, fails := check(tc, stepNumber, assertion, executorResult)
+		errs := check(ctx, tc, stepNumber, rangedIndex, assertion, executorResult)
+		isAssertionOK := true
 		if errs != nil {
 			errors = append(errors, *errs)
 			isOK = false
+			isAssertionOK = false
 		}
-		if fails != nil {
-			failures = append(failures, *fails)
-			isOK = false
-		}
+		assertions = append(assertions, AssertionApplied{
+			Assertion: assertion,
+			IsOK:      isAssertionOK,
+		})
 	}
 
 	if _, ok := executorResult["result.systemerr"]; ok {
@@ -67,13 +72,20 @@ func applyAssertions(r interface{}, tc TestCase, stepNumber int, step TestStep, 
 		systemout = fmt.Sprintf("%v", executorResult["result.systemout"])
 	}
 
-	return assertionsApplied{isOK, errors, failures, systemout, systemerr}
+	return AssertionsApplied{
+		OK:         isOK,
+		errors:     errors,
+		systemerr:  systemerr,
+		systemout:  systemout,
+		Assertions: assertions,
+	}
 }
 
 type assertion struct {
-	Actual interface{}
-	Func   assertions.AssertFunc
-	Args   []interface{}
+	Actual   interface{}
+	Func     assertions.AssertFunc
+	Args     []interface{}
+	Required bool
 }
 
 func parseAssertions(ctx context.Context, s string, input interface{}) (*assertion, error) {
@@ -83,6 +95,13 @@ func parseAssertions(ctx context.Context, s string, input interface{}) (*asserti
 		return nil, errors.New("assertion syntax error")
 	}
 	actual := dump[assert[0]]
+
+	// "Must" assertions use same tests as "Should" ones, only the flag changes
+	required := false
+	if strings.HasPrefix(assert[1], "Must") {
+		required = true
+		assert[1] = strings.Replace(assert[1], "Must", "Should", 1)
+	}
 
 	f, ok := assertions.Get(assert[1])
 	if !ok {
@@ -97,22 +116,110 @@ func parseAssertions(ctx context.Context, s string, input interface{}) (*asserti
 		}
 	}
 	return &assertion{
-		Actual: actual,
-		Func:   f,
-		Args:   args,
+		Actual:   actual,
+		Func:     f,
+		Args:     args,
+		Required: required,
 	}, nil
 }
 
-func check(tc TestCase, stepNumber int, assertion string, r interface{}) (*Failure, *Failure) {
+// check selects the correct assertion function to call depending on typing provided by user
+func check(ctx context.Context, tc TestCase, stepNumber int, rangedIndex int, assertion Assertion, r interface{}) *Failure {
+	var errs *Failure
+	switch t := assertion.(type) {
+	case string:
+		errs = checkString(ctx, tc, stepNumber, rangedIndex, assertion.(string), r)
+	case map[string]interface{}:
+		errs = checkBranch(ctx, tc, stepNumber, rangedIndex, assertion.(map[string]interface{}), r)
+	default:
+		errs = newFailure(ctx, tc, stepNumber, rangedIndex, "", fmt.Errorf("unsupported assertion format: %v", t))
+	}
+	return errs
+}
+
+// checkString evaluate a complex assertion containing logical operators
+// it recursively calls checkAssertion for each operand
+func checkBranch(ctx context.Context, tc TestCase, stepNumber int, rangedIndex int, branch map[string]interface{}, r interface{}) *Failure {
+	// Extract logical operator
+	if len(branch) != 1 {
+		return newFailure(ctx, tc, stepNumber, rangedIndex, "", fmt.Errorf("expected exactly 1 logical operator but %d were provided", len(branch)))
+	}
+	var operator string
+	for k := range branch {
+		operator = k
+	}
+
+	// Extract logical operands
+	var operands []interface{}
+	switch t := branch[operator].(type) {
+	case []interface{}:
+		operands = branch[operator].([]interface{})
+	default:
+		return newFailure(ctx, tc, stepNumber, rangedIndex, "", fmt.Errorf("expected %s operands to be an []interface{}, got %v", operator, t))
+	}
+	if len(operands) == 0 {
+		return nil
+	}
+
+	// Evaluate assertions (operands)
+	var results []string
+	assertionsCount := len(operands)
+	assertionsSuccess := 0
+	for _, assertion := range operands {
+		errs := check(ctx, tc, stepNumber, rangedIndex, assertion, r)
+		if errs != nil {
+			results = append(results, fmt.Sprintf("  - fail: %s", assertion))
+		}
+		if errs == nil {
+			assertionsSuccess++
+			results = append(results, fmt.Sprintf("  - pass: %s", assertion))
+		}
+	}
+
+	// Evaluate operator behaviour
+	var err error
+	switch operator {
+	case "and":
+		if assertionsSuccess != assertionsCount {
+			err = fmt.Errorf("%d/%d assertions succeeded:\n%s\n", assertionsSuccess, assertionsCount, strings.Join(results, "\n"))
+		}
+	case "or":
+		if assertionsSuccess == 0 {
+			err = fmt.Errorf("no assertions succeeded:\n%s\n", strings.Join(results, "\n"))
+		}
+	case "xor":
+		if assertionsSuccess == 0 {
+			err = fmt.Errorf("no assertions succeeded:\n%s\n", strings.Join(results, "\n"))
+		}
+		if assertionsSuccess > 1 {
+			err = fmt.Errorf("multiple assertions succeeded but expected only one to succeed:\n%s\n", strings.Join(results, "\n"))
+		}
+	case "not":
+		if assertionsSuccess > 0 {
+			err = fmt.Errorf("some assertions succeeded but expected none to succeed:\n%s\n", strings.Join(results, "\n"))
+		}
+	default:
+		return newFailure(ctx, tc, stepNumber, rangedIndex, "", fmt.Errorf("unsupported assertion operator %s", operator))
+	}
+	if err != nil {
+		return newFailure(ctx, tc, stepNumber, rangedIndex, "", err)
+	}
+	return nil
+}
+
+// checkString evaluate a single string assertion
+func checkString(ctx context.Context, tc TestCase, stepNumber int, rangedIndex int, assertion string, r interface{}) *Failure {
 	assert, err := parseAssertions(context.Background(), assertion, r)
 	if err != nil {
-		return nil, newFailure(tc, stepNumber, assertion, err)
+		return newFailure(ctx, tc, stepNumber, rangedIndex, assertion, err)
 	}
 
 	if err := assert.Func(assert.Actual, assert.Args...); err != nil {
-		return nil, newFailure(tc, stepNumber, assertion, err)
+		failure := newFailure(ctx, tc, stepNumber, rangedIndex, assertion, err)
+		failure.AssertionRequired = assert.Required
+		return failure
 	}
-	return nil, nil
+	return nil
 }
 
 // splitAssertion splits the assertion string a, with support
@@ -246,4 +353,22 @@ func findLineNumber(filename, testcase string, stepNumber int, assertion string,
 	}
 
 	return countLine
+}
+
+// This evaluates a string of assertions with a given vars scope, and returns a slice of failures (i.e. empty slice = all pass)
+func testConditionalStatement(ctx context.Context, tc *TestCase, assertions []string, vars H, text string) ([]string, error) {
+	var failures []string
+	for _, assertion := range assertions {
+		Debug(ctx, "evaluating %s", assertion)
+		assert, err := parseAssertions(ctx, assertion, vars)
+		if err != nil {
+			Error(ctx, "unable to parse assertion: %v", err)
+			return failures, err
+		}
+		if err := assert.Func(assert.Actual, assert.Args...); err != nil {
+			s := fmt.Sprintf(text, tc.originalName, err)
+			failures = append(failures, s)
+		}
+	}
+	return failures, nil
 }
