@@ -3,10 +3,12 @@ package cmd
 import (
 	"fmt"
 	"os"
+
+	"github.com/ovh/venom"
 )
 
 // Exit func display an error message on stderr and exit 1
 func Exit(format string, args ...interface{}) {
 	fmt.Fprintf(os.Stderr, format, args...)
-	os.Exit(1)
+	venom.OSExit(1)
 }
