@@ -9,7 +9,7 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-func TestProcessVariableAssigments(t *testing.T) {
+func TestProcessVariableAssignments(t *testing.T) {
 	InitTestLogger(t)
 	assign := AssignStep{}
 	assign.Assignments = make(map[string]Assignment)
@@ -26,7 +26,7 @@ func TestProcessVariableAssigments(t *testing.T) {
 
 	tcVars := H{"here.some.value": "this is the \nvalue"}
 
-	result, is, err := processVariableAssigments(context.TODO(), "", tcVars, b)
+	result, is, err := processVariableAssignments(context.TODO(), "", tcVars, b)
 	assert.True(t, is)
 	assert.NoError(t, err)
 	assert.NotNil(t, result)
@@ -38,7 +38,7 @@ func TestProcessVariableAssigments(t *testing.T) {
 script: echo 'foo'
 `)
 	assert.NoError(t, yaml.Unmarshal(b, &wrongStepIn))
-	result, is, err = processVariableAssigments(context.TODO(), "", tcVars, b)
+	result, is, err = processVariableAssignments(context.TODO(), "", tcVars, b)
 	assert.False(t, is)
 	assert.NoError(t, err)
 	assert.Nil(t, result)
