@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"os"
 	"runtime/pprof"
+	"time"
 
-	"github.com/fatih/color"
 	"github.com/gosimple/slug"
 	"github.com/ovh/cds/sdk/interpolate"
 	log "github.com/sirupsen/logrus"
@@ -32,7 +32,7 @@ func (v *Venom) runTestSuite(ctx context.Context, ts *TestSuite) {
 		}
 	}
 
-	// Intialiaze the testsuite varibles and compute a first interpolation over them
+	// Intialiaze the testsuite variables and compute a first interpolation over them
 	ts.Vars.AddAll(v.variables.Clone())
 	vars, _ := DumpStringPreserveCase(ts.Vars)
 	for k, v := range vars {
@@ -43,6 +43,15 @@ func (v *Venom) runTestSuite(ctx context.Context, ts *TestSuite) {
 		ts.Vars.Add(k, computedV)
 	}
 
+	exePath, err := os.Executable()
+	if err != nil {
+		log.Errorf("failed to get executable path: %v", err)
+	} else {
+		ts.Vars.Add("venom.executable", exePath)
+	}
+
+	ts.Vars.Add("venom.outputdir", v.OutputDir)
+	ts.Vars.Add("venom.libdir", v.LibDir)
 	ts.Vars.Add("venom.testsuite", ts.Name)
 	ts.ComputedVars = H{}
 
@@ -55,82 +64,140 @@ func (v *Venom) runTestSuite(ctx context.Context, ts *TestSuite) {
 		totalSteps += len(tc.testSteps)
 	}
 
+	ts.Status = StatusRun
+
+	// ##### RUN Test Cases Here
 	v.runTestCases(ctx, ts)
+
+	var isFailed bool
+	var nSkip int
+	for _, tc := range ts.TestCases {
+		if tc.Status == StatusFail {
+			isFailed = true
+			ts.NbTestcasesFail++
+		} else if tc.Status == StatusSkip {
+			nSkip++
+			ts.NbTestcasesSkip++
+		} else if tc.Status == StatusPass {
+			ts.NbTestcasesPass++
+		}
+	}
+
+	if isFailed {
+		ts.Status = StatusFail
+		v.Tests.NbTestsuitesFail++
+	} else if nSkip > 0 && nSkip == len(ts.TestCases) {
+		ts.Status = StatusSkip
+		v.Tests.NbTestsuitesSkip++
+	} else {
+		ts.Status = StatusPass
+		v.Tests.NbTestsuitesPass++
+	}
 }
 
 func (v *Venom) runTestCases(ctx context.Context, ts *TestSuite) {
-	var red = color.New(color.FgRed).SprintFunc()
-	var green = color.New(color.FgGreen).SprintFunc()
-	var cyan = color.New(color.FgCyan).SprintFunc()
-	var gray = color.New(color.Attribute(90)).SprintFunc()
+	verboseReport := v.Verbose >= 1
 
-	v.Println(" • %s (%s)", ts.Name, ts.Package)
+	v.Println(" • %s (%s)", ts.Name, ts.Filepath)
 
 	for i := range ts.TestCases {
 		tc := &ts.TestCases[i]
+		tc.IsEvaluated = true
 		v.Print(" \t• %s", tc.Name)
-		tc.Classname = ts.Filename
 		var hasFailure bool
+		var hasRanged bool
 		var hasSkipped = len(tc.Skipped) > 0
 		if !hasSkipped {
+			start := time.Now()
+			tc.Start = start
+			ts.Status = StatusRun
+			if verboseReport || hasRanged {
+				v.Print("\n")
+			}
+			// ##### RUN Test Case Here
 			v.runTestCase(ctx, ts, tc)
+			tc.End = time.Now()
+			tc.Duration = tc.End.Sub(tc.Start).Seconds()
 		}
 
-		if len(tc.Failures) > 0 {
-			ts.Failures += len(tc.Failures)
-			hasFailure = true
-		}
-		if len(tc.Errors) > 0 {
-			ts.Errors += len(tc.Errors)
-			hasFailure = true
-		}
-		if len(tc.Skipped) > 0 {
-			ts.Skipped += len(tc.Skipped)
-			hasSkipped = true
-		}
-
-		if hasSkipped {
-			v.Println(" %s", gray("SKIPPED"))
-			continue
+		skippedSteps := 0
+		for _, testStepResult := range tc.TestStepResults {
+			if testStepResult.RangedEnable {
+				hasRanged = true
+			}
+			if testStepResult.Status == StatusFail {
+				hasFailure = true
+			}
+			if testStepResult.Status == StatusSkip {
+				skippedSteps++
+			}
 		}
 
 		if hasFailure {
-			v.Println(" %s", red("FAILURE"))
+			tc.Status = StatusFail
+		} else if skippedSteps == len(tc.TestStepResults) {
+			//If all test steps were skipped, consider the test case as skipped
+			tc.Status = StatusSkip
+		} else if tc.Status != StatusSkip {
+			tc.Status = StatusPass
+		}
+
+		// Verbose mode already reported tests status, so just print them when non-verbose
+		indent := ""
+		if hasRanged || verboseReport {
+			indent = "\t  "
+			// If the testcase was entirely skipped, then the verbose mode will not have any output
+			// Print something to inform that the testcase was indeed processed although skipped
+			if len(tc.TestStepResults) == 0 {
+				v.Println("\t\t%s", Gray("• (all steps were skipped)"))
+				continue
+			}
 		} else {
-			v.Println(" %s", green("SUCCESS"))
+			if hasFailure {
+				v.Println(" %s", Red(StatusFail))
+			} else if tc.Status == StatusSkip {
+				v.Println(" %s", Gray(StatusSkip))
+				continue
+			} else {
+				v.Println(" %s", Green(StatusPass))
+			}
 		}
 
 		for _, i := range tc.computedInfo {
-			v.Println("\t  %s %s", cyan("[info]"), cyan(i))
+			v.Println("\t  %s%s %s", indent, Cyan("[info]"), Cyan(i))
 		}
 
 		for _, i := range tc.computedVerbose {
-			v.PrintlnTrace(i)
+			v.PrintlnIndentedTrace(i, indent)
 		}
 
-		if hasFailure {
-			for _, f := range tc.Failures {
-				v.Println("%s", red(f.Value))
-			}
-			for _, f := range tc.Errors {
-				v.Println("%s", red(f.Value))
+		// Verbose mode already reported failures, so just print them when non-verbose
+		if !hasRanged && !verboseReport && hasFailure {
+			for _, testStepResult := range tc.TestStepResults {
+				for _, f := range testStepResult.Errors {
+					v.Println("%s", Yellow(f.Value))
+				}
 			}
 		}
 
-		if v.StopOnFailure && (len(tc.Failures) > 0 || len(tc.Errors) > 0) {
-			// break TestSuite
-			return
+		if v.StopOnFailure {
+			for _, testStepResult := range tc.TestStepResults {
+				if len(testStepResult.Errors) > 0 {
+					// break TestSuite
+					return
+				}
+			}
 		}
 		ts.ComputedVars.AddAllWithPrefix(tc.Name, tc.computedVars)
 	}
 }
 
-//Parse the suite to find unreplaced and extracted variables
+// Parse the suite to find unreplaced and extracted variables
 func (v *Venom) parseTestSuite(ts *TestSuite) ([]string, []string, error) {
 	return v.parseTestCases(ts)
 }
 
-//Parse the testscases to find unreplaced and extracted variables
+// Parse the testscases to find unreplaced and extracted variables
 func (v *Venom) parseTestCases(ts *TestSuite) ([]string, []string, error) {
 	var vars []string
 	var extractsVars []string
