@@ -12,7 +12,7 @@ func Test_RemoveNotPrintableChar(t *testing.T) {
 		want string
 	}{
 		{
-			name: "remove U+001B espace code (not printable)",
+			name: "remove U+001B escape code (not printable)",
 			// line below contains escapce code U+001B
 			args: args{in: "python-mysqldb : [34mOK[0m"},
 			want: "python-mysqldb :  [34mOK [0m",
