@@ -5,11 +5,10 @@ import (
 	"encoding/json"
 	"encoding/xml"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
+	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/fatih/color"
 	tap "github.com/mndrix/tap-go"
@@ -18,47 +17,97 @@ import (
 )
 
 func init() {
-	if strings.ToLower(os.Getenv("IS_TTY")) == "true" || os.Getenv("IS_TTY") == "1" {
+	color.NoColor = true
+	if os.Getenv("IS_TTY") == "" || strings.ToLower(os.Getenv("IS_TTY")) == "true" || os.Getenv("IS_TTY") == "1" {
 		color.NoColor = false
 	}
 }
 
 // OutputResult output result to sdtout, files...
-func (v *Venom) OutputResult(tests Tests, elapsed time.Duration) error {
+func (v *Venom) OutputResult() error {
 	if v.OutputDir == "" {
 		return nil
 	}
-	var data []byte
-	var err error
-	switch v.OutputFormat {
-	case "json":
-		data, err = json.MarshalIndent(tests, "", "  ")
-		if err != nil {
-			log.Fatalf("Error: cannot format output json (%s)", err)
+
+	for i := range v.Tests.TestSuites {
+		tcFiltered := []TestCase{}
+		for _, tc := range v.Tests.TestSuites[i].TestCases {
+			if tc.IsEvaluated {
+				tcFiltered = append(tcFiltered, tc)
+			}
 		}
-	case "tap":
-		data, err = outputTapFormat(tests)
-		if err != nil {
-			log.Fatalf("Error: cannot format output tap (%s)", err)
+		v.Tests.TestSuites[i].TestCases = tcFiltered
+
+		testsResult := &Tests{
+			TestSuites:       []TestSuite{v.Tests.TestSuites[i]},
+			Status:           v.Tests.Status,
+			NbTestsuitesFail: v.Tests.NbTestsuitesFail,
+			NbTestsuitesPass: v.Tests.NbTestsuitesPass,
+			NbTestsuitesSkip: v.Tests.NbTestsuitesSkip,
+			Duration:         v.Tests.Duration,
+			Start:            v.Tests.Start,
+			End:              v.Tests.End,
 		}
-	case "yml", "yaml":
-		data, err = yaml.Marshal(tests)
-		if err != nil {
-			log.Fatalf("Error: cannot format output yaml (%s)", err)
+
+		var data []byte
+		var err error
+
+		switch v.OutputFormat {
+		case "json":
+			data, err = json.MarshalIndent(testsResult, "", "  ")
+			if err != nil {
+				log.Fatalf("Error: cannot format output json (%s)", err)
+			}
+		case "tap":
+			data, err = outputTapFormat(*testsResult)
+			if err != nil {
+				log.Fatalf("Error: cannot format output tap (%s)", err)
+			}
+		case "yml", "yaml":
+			data, err = yaml.Marshal(testsResult)
+			if err != nil {
+				log.Fatalf("Error: cannot format output yaml (%s)", err)
+			}
+		case "xml":
+			data, err = outputXMLFormat(*testsResult)
+			if err != nil {
+				log.Fatalf("Error: cannot format output xml (%s)", err)
+			}
+		case "html":
+			log.Fatalf("Error: you have to use the --html-report flag")
 		}
-	default:
-		dataxml, errm := xml.MarshalIndent(tests, "", "  ")
-		if errm != nil {
-			log.Fatalf("Error: cannot format xml output: %s", errm)
+
+		fname := strings.TrimSuffix(v.Tests.TestSuites[i].Filepath, filepath.Ext(v.Tests.TestSuites[i].Filepath))
+		fname = strings.ReplaceAll(fname, "/", "_")
+		filename := path.Join(v.OutputDir, "test_results_"+fname+"."+v.OutputFormat)
+		if err := os.WriteFile(filename, data, 0600); err != nil {
+			return fmt.Errorf("Error while creating file %s: %v", filename, err)
 		}
-		data = append([]byte(`<?xml version="1.0" encoding="utf-8"?>`), dataxml...)
+		v.PrintFunc("Writing file %s\n", filename)
 	}
 
-	filename := path.Join(v.OutputDir, "test_results."+v.OutputFormat)
-	if err := ioutil.WriteFile(filename, data, 0600); err != nil {
-		return fmt.Errorf("Error while creating file %s: %v", filename, err)
+	if v.HtmlReport {
+		testsResult := &Tests{
+			TestSuites:       v.Tests.TestSuites,
+			Status:           v.Tests.Status,
+			NbTestsuitesFail: v.Tests.NbTestsuitesFail,
+			NbTestsuitesPass: v.Tests.NbTestsuitesPass,
+			NbTestsuitesSkip: v.Tests.NbTestsuitesSkip,
+			Duration:         v.Tests.Duration,
+			Start:            v.Tests.Start,
+			End:              v.Tests.End,
+		}
+
+		data, err := outputHTML(testsResult)
+		if err != nil {
+			log.Fatalf("Error: cannot format output html (%s)", err)
+		}
+		var filename = filepath.Join(v.OutputDir, computeOutputFilename("test_results.html"))
+		v.PrintFunc("Writing html file %s\n", filename)
+		if err := os.WriteFile(filename, data, 0600); err != nil {
+			return fmt.Errorf("Error while creating file %s: %v", filename, err)
+		}
 	}
-	v.PrintFunc("Writing file %s\n", filename)
 
 	return nil
 }
@@ -67,34 +116,86 @@ func outputTapFormat(tests Tests) ([]byte, error) {
 	tapValue := tap.New()
 	buf := new(bytes.Buffer)
 	tapValue.Writer = buf
-	tapValue.Header(tests.Total)
+	var total int
 	for _, ts := range tests.TestSuites {
 		for _, tc := range ts.TestCases {
+			total++
 			name := ts.Name + " / " + tc.Name
 			if len(tc.Skipped) > 0 {
 				tapValue.Skip(1, name)
 				continue
 			}
 
-			if len(tc.Errors) > 0 {
-				tapValue.Fail(name)
-				for _, e := range tc.Errors {
-					tapValue.Diagnosticf("Error: %s", e.Value)
+			for _, testStepResult := range tc.TestStepResults {
+				if len(testStepResult.Errors) > 0 {
+					tapValue.Fail(name)
+					for _, e := range testStepResult.Errors {
+						tapValue.Diagnosticf("Error: %s", e.Value)
+					}
+					continue
 				}
-				continue
-			}
-
-			if len(tc.Failures) > 0 {
-				tapValue.Fail(name)
-				for _, e := range tc.Failures {
-					tapValue.Diagnosticf("Failure: %s", e.Value)
-				}
-				continue
 			}
 
 			tapValue.Pass(name)
 		}
 	}
+	tapValue.Header(total)
 
 	return buf.Bytes(), nil
+}
+
+func outputXMLFormat(tests Tests) ([]byte, error) {
+
+	testsXML := TestsXML{}
+
+	for _, ts := range tests.TestSuites {
+		tsXML := TestSuiteXML{
+			Name:    ts.Name,
+			Package: ts.Filepath,
+			Time:    fmt.Sprintf("%f", ts.Duration),
+		}
+
+		for _, tc := range ts.TestCases {
+			switch tc.Status {
+			case StatusFail:
+				tsXML.Errors++
+			case StatusSkip:
+				tsXML.Skipped++
+			}
+			tsXML.Total++
+
+			failuresXML := []FailureXML{}
+			systemout := InnerResult{}
+			systemerr := InnerResult{}
+			for _, result := range tc.TestStepResults {
+				for _, failure := range result.Errors {
+					failuresXML = append(failuresXML, FailureXML{
+						Value: failure.Value,
+					})
+				}
+				systemout.Value += result.Systemout
+				systemerr.Value += result.Systemerr
+			}
+
+			tcXML := TestCaseXML{
+				Classname: ts.Filename,
+				Errors:    failuresXML,
+				Name:      tc.Name,
+				Skipped:   tc.Skipped,
+				Systemout: systemout,
+				Systemerr: systemerr,
+				Time:      tc.Duration,
+			}
+			tsXML.TestCases = append(tsXML.TestCases, tcXML)
+		}
+		testsXML.TestSuites = append(testsXML.TestSuites, tsXML)
+	}
+
+	dataxml, errm := xml.MarshalIndent(testsXML, "", "  ")
+	if errm != nil {
+		log.Fatalf("Error: cannot format xml output: %s", errm)
+	}
+	data := append([]byte(`<?xml version="1.0" encoding="utf-8"?>`), dataxml...)
+
+	return data, nil
 }
